+++ domain/employee_shift_dto.go
@@ -7,9 +7,9 @@ type EmployeeShiftResponse struct {
 	BranchUUID string                           `json:"branch_uuid"`
 	UserUUID   string                           `json:"user_uuid"`
 	StartTime  time.Time                        `json:"start_time"`
-	StartCash  float32                          `json:"start_cash"`
+	StartCash  float64                          `json:"start_cash"`
 	EndTime    *time.Time                       `json:"end_time"`
-	EndCash    *float32                         `json:"end_cash"`
+	EndCash    *float64                         `json:"end_cash"`
 	Supporters []EmployeeShiftSupporterResponse `json:"supporters"`
 	CreatedAt  time.Time                        `json:"created_at"`
 	UpdatedAt  *time.Time                       `json:"updated_at"`
@@ -30,11 +30,11 @@ type EmployeeShiftClockInRequest struct {
 	UserUUID       string   `json:"user_uuid"`
 	SupportingUUID *string  `json:"supporting_uuid"`
 	StartTime      string   `json:"start_time"`
-	StartCash      *float32 `json:"start_cash"`
+	StartCash      *float64 `json:"start_cash"`
 }
 
 type EmployeeShiftClockOutRequest struct {
 	UUID    string   `json:"uuid"`
 	EndTime string   `json:"end_time"`
-	EndCash *float32 `json:"end_cash"`
+	EndCash *float64 `json:"end_cash"`
 }
